requirements.txt
# This file may be used to create an environment using:
# $ conda create --name <env> --file <this file>
# platform: win-64
altair=4.2.2=pypi_0
asgiref=3.6.0=pypi_0
attrs=23.1.0=pypi_0
backports-zoneinfo=0.2.1=pypi_0
blinker=1.6.2=pypi_0
bzip2=1.0.8=h8ffe710_4
ca-certificates=2023.5.7=h56e8100_0
cachetools=5.3.0=pypi_0
certifi=2023.5.7=pypi_0
charset-normalizer=3.1.0=pypi_0
click=8.1.3=pypi_0
colorama=0.4.6=pypi_0
decorator=5.1.1=pypi_0
django=3.2.19=pypi_0
eigen=3.4.0=h2d74725_0
entrypoints=0.4=pypi_0
ffmpeg=4.2.3=ha925a31_0
freetype=2.12.1=h546665d_1
gettext=0.21.1=h5728263_0
gitdb=4.0.10=pypi_0
gitpython=3.1.31=pypi_0
glib=2.76.2=h12be248_0
glib-tools=2.76.2=h12be248_0
gst-plugins-base=1.18.5=h9e645db_0
gstreamer=1.18.5=hd78058f_0
hdf5=1.10.6=nompi_h5268f04_1114
icu=58.2=ha925a31_3
idna=3.4=pypi_0
importlib-metadata=6.6.0=pypi_0
importlib-resources=5.12.0=pypi_0
intel-openmp=2023.1.0=h57928b3_46319
jinja2=3.1.2=pypi_0
jpeg=9e=hcfcfb64_3
jsonschema=4.17.3=pypi_0
krb5=1.20.1=h6609f42_0
lerc=4.0.0=h63175ca_0
libblas=3.9.0=16_win64_mkl
libcblas=3.9.0=16_win64_mkl
libclang=12.0.0=default_h627e005_2
libcurl=8.1.0=h68f0423_0
libdeflate=1.17=hcfcfb64_0
libffi=3.4.2=h8ffe710_5
libglib=2.76.2=he8f3873_0
libhwloc=2.9.1=h51c2c0f_0
libiconv=1.17=h8ffe710_0
liblapack=3.9.0=16_win64_mkl
libogg=1.3.4=h8ffe710_1
libpng=1.6.39=h19919ed_0
libprotobuf=3.20.3=h12be248_0
libsqlite=3.42.0=hcfcfb64_0
libssh2=1.10.0=h680486a_3
libtiff=4.5.0=hf8721a0_2
libvorbis=1.3.7=h0e60522_0
libwebp=1.3.0=hcfcfb64_0
libwebp-base=1.3.0=hcfcfb64_0
libxml2=2.10.4=hc3477c8_0
libxslt=1.1.37=h0192164_0
libzlib=1.2.13=hcfcfb64_4
llvmlite=0.39.1=pypi_0
markdown-it-py=2.2.0=pypi_0
markupsafe=2.1.2=pypi_0
mdurl=0.1.2=pypi_0
mkl=2022.1.0=h6a75c08_874
numba=0.56.4=pypi_0
numpy=1.21.6=py37h2830a78_0
opencv=4.6.0=py37hf11a4ad_3
opencv-python=4.7.0.72=pypi_0
openssl=1.1.1t=hcfcfb64_0
packaging=23.1=pypi_0
pandas=1.3.5=pypi_0
pcre2=10.40=h17e33f8_0
pillow=9.5.0=pypi_0
pip=23.1.2=pyhd8ed1ab_0
pkgutil-resolve-name=1.3.10=pypi_0
protobuf=3.20.3=pypi_0
pthreads-win32=2.9.1=hfa6e2cd_3
pyarrow=12.0.0=pypi_0
pydeck=0.8.1b0=pypi_0
pygments=2.15.1=pypi_0
pympler=1.0.1=pypi_0
pyrsistent=0.19.3=pypi_0
python=3.7.12=h7840368_100_cpython
python-dateutil=2.8.2=pypi_0
python_abi=3.7=3_cp37m
pytz=2023.3=pypi_0
qt-main=5.15.2=he8e5bd7_7
qt-webengine=5.15.9=hb9a9bb5_5
qtwebkit=5.212=h2bbfb41_5
requests=2.30.0=pypi_0
rich=13.3.5=pypi_0
setuptools=67.7.2=pyhd8ed1ab_0
six=1.16.0=pypi_0
smmap=5.0.0=pypi_0
sqlite=3.42.0=hcfcfb64_0
sqlparse=0.4.4=pypi_0
streamlit=1.22.0=pypi_0
tbb=2021.9.0=h91493d7_0
tenacity=8.2.2=pypi_0
toml=0.10.2=pypi_0
toolz=0.12.0=pypi_0
tornado=6.2=pypi_0
typing-extensions=4.5.0=pypi_0
tzdata=2023.3=pypi_0
tzlocal=5.0.1=pypi_0
ucrt=10.0.22621.0=h57928b3_0
urllib3=2.0.2=pypi_0
validators=0.20.0=pypi_0
vc=14.3=hb25d44b_16
vc14_runtime=14.34.31931=h5081d32_16
vs2015_runtime=14.34.31931=hed1258a_16
watchdog=3.0.0=pypi_0
wheel=0.40.0=pyhd8ed1ab_0
xz=5.2.6=h8d14728_0
zipp=3.15.0=pypi_0
zlib=1.2.13=hcfcfb64_4
zstd=1.5.2=h12be248_6
